# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SERVED. On-Demand Process Serving Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List all process servers
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Create a process server profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerBody"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get a server profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    patch:
      operationId: updateServer
      tags: [servers]
      summary: Update a server profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerBody"
      responses:
        "200":
          description: Updated server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
...
  /storage/server-photo/request-url:
    post:
      tags: [storage]
      operationId: requestServerPhotoUploadUrl
      summary: Request a presigned URL for a server proof photo
      description: |
        Server-only. Returns a presigned upload URL for a GPS-stamped photo
        captured during a service attempt or successful service. Caller must
        be the assigned server on the job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerPhotoUploadRequest'
      responses:
        '200':
          description: Presigned upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPhotoUploadResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Caller is not the assigned server on the job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/profile-photo/request-url:
    post:
      tags: [storage]
      operationId: requestProfilePhotoUploadUrl
      summary: Request a presigned URL for a server profile headshot
      description: |
        Server-only. Returns a presigned upload URL for the caller's own
        profile headshot. No job binding — any authenticated server may
        request one for themselves.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfilePhotoUploadRequest'
      responses:
        '200':
          description: Presigned upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerPhotoUploadResponse'
        '403':
          description: Caller is not a server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '413':
          description: Upload would exceed the attorney's plan storage quota.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    ServerPhotoUploadRequest:
      type: object
      properties:
        jobId:
          type: integer
        contentType:
          type: string
        size:
          type: integer
      required:
        - jobId
        - contentType
        - size
    ServerPhotoUploadResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath
    Server:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serverTier:
          type: string
          enum: [basic, pro, licensed]
        licenseNumber:
          type: string
        serviceArea:
          type: string
        businessAddress:
          type: string
          nullable: true
          description: Server's business / mailing address. Rendered into the affidavit attestation block.
        isLicensedNvServer:
          type: boolean
          description: Server attests they hold a Nevada PILB Process Server work card.
        licenseCounty:
          type: string
          nullable: true
          description: Nevada county of license registration (when isLicensedNvServer = true).
        serverType:
          type: string
          nullable: true
          enum: [licensed_nv, registered, private, sheriff]
          description: Self-declared server classification rendered on the affidavit attestation block.
        photoUrl:
          type: string
          nullable: true
          description: Object-storage path of the server's profile headshot (from /storage/profile-photo/request-url). Null until uploaded.
        active:
          type: boolean
        jobsCompleted:
          type: integer
        stripeAccountId:
          type: string
          nullable: true
          description: Stripe Connect Express account id (null until onboarding starts).
        payoutsEnabled:
          type: boolean
          description: True iff Stripe reports both charges_enabled && payouts_enabled.
        verifiedAt:
          type: string
          format: date-time
          nullable: true
          description: Mirrored from server_credentials.verified_at; populated after the background check passes.
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - serverTier
        - active
        - jobsCompleted
        - payoutsEnabled
        - createdAt
    CreateServerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        serverTier:
          type: string
          enum: [basic, pro, licensed]
        licenseNumber:
          type: string
        serviceArea:
          type: string
        businessAddress:
          type: string
        isLicensedNvServer:
          type: boolean
        licenseCounty:
          type: string
      required:
        - name
        - email
        - serverTier
    UpdateServerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        serverTier:
          type: string
          enum: [basic, pro, licensed]
        licenseNumber:
          type: string
        serviceArea:
          type: string
        active:
          type: boolean
        businessAddress:
          type: string
        isLicensedNvServer:
          type: boolean
        licenseCounty:
          type: string
...
    ProfilePhotoUploadRequest:
      type: object
      description: Optional metadata for a server profile-photo upload. No job binding.
      properties:
        contentType:
          type: string
        size:
          type: integer
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `application/pdf`).
        jobId:
          type: integer
          description: Optional job to associate the upload with.
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Persist this.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...